增加ping测试,本来如果阿里云和外网(搬瓦工)速度够快,我就在中间搭一层代理中转,结果tm比我本地还慢...

+++ LiteDbLog/LiteDBLog/DBLogDao.cs
@@ -26,7 +26,7 @@ namespace LiteDbLog.LiteDBLog
                 Name = _cnName,
                 Content = content,
                 Title = title,
-                Level = level
+                Level = (int)level
             };
             Insert(en);
         }
@@ -38,7 +38,7 @@ namespace LiteDbLog.LiteDBLog
                 Name = _cnName,
                 Content = content,
                 Title = title,
-                Level = level,
+                Level = (int)level,
                 Exception = ex
             };
             Insert(en);
